A Sexp Array is encoded and put in the tuplespace for EvalC The problem is that, by now, I am not able to decoade it in the agent

+++ agent/main.go
@@ -1,21 +1,17 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/gob"
 	"github.com/coreos/etcd/clientv3"
 	"github.com/ditrit/go-linda"
 	zygo "github.com/glycerine/zygomys/repl"
-	//"github.com/google/uuid"
+	"github.com/google/uuid"
 	"github.com/kelseyhightower/envconfig"
 	"log"
 	"math/rand"
 	"os"
 	"time"
 )
-
-const prefix = "LINDA"
 
 type configuration struct {
 	EtcdEndpoints []string      `envconfig:"etcd_endpoint" default:"localhost:2379,localhost:22379,localhost:32379"`
@@ -29,6 +25,7 @@ func sleep(env *zygo.Glisp, name string, args []zygo.Sexp) (zygo.Sexp, error) {
 }
 
 func main() {
+	me := uuid.New()
 	var s configuration
 	err := envconfig.Process("glinda", &s)
 	if err != nil {
@@ -48,46 +45,56 @@ func main() {
 		log.Fatalf("Cannot connect to etcd: %v", err)
 	}
 	defer cli.Close()
+	// Getting the lisp source from ETCD
+	ctx, cancel := context.WithTimeout(context.Background(), s.Timeout)
+	resp, err := cli.Get(ctx, os.Args[1])
+	cancel()
+	if err != nil {
+		log.Fatal("Cannot get lisp code from the tuple space (etcd)", err)
+	}
+	if len(resp.Kvs) != 1 {
+		log.Fatalf("Found %v lisp code in the tuple space mtching ID %v", len(resp.Kvs), os.Args[1])
+	}
+	var lisp []byte
+	for _, ev := range resp.Kvs {
+		lisp = ev.Value
+		//fmt.Printf("%s : %s\n", ev.Key, ev.Value)
+	}
 
+	lda := linda.New(cli)
+	env := zygo.NewGlisp()
+	env.AddFunction("in", lda.InRd)
+	env.AddFunction("rd", lda.InRd)
+	env.AddFunction("out", lda.Out)
+	env.AddFunction("eval", lda.Eval)
+	env.AddFunction("sleep", sleep)
+	//env.SourceFile(f)
+	env.LoadString(string(lisp))
 	// The tuple does not exists, watch for a new event until a tuple match
-	var msg linda.Message
 	rch := cli.Watch(context.Background(), "LINDA-evalc-", clientv3.WithPrefix())
-Loop:
 	for wresp := range rch {
 		for _, ev := range wresp.Events {
-			// TODO: Check if ev.Type is PUT otherwise continue
-			//fmt.Printf("%s %q : %q\n", ev.Type, ev.Kv.Key, ev.Kv.Value)
-			// TODO: Remove the tuple from the space is IN is called
-			_, err := cli.Delete(context.TODO(), string(ev.Kv.Key), clientv3.WithPrefix())
-			if err != nil {
-				log.Println(err)
-				continue
+			// Check if the evalc is anonymous or for me...
+			if string(ev.Kv.Key) == "LINDA-evalc-" || string(ev.Kv.Key) == "LINDA-evalc-"+me.String() {
+
+				// TODO: Check if ev.Type is PUT otherwise continue
+				//fmt.Printf("%s %q : %q\n", ev.Type, ev.Kv.Key, ev.Kv.Value)
+				// TODO: Remove the tuple from the space is IN is called
+				_, err := cli.Delete(context.TODO(), string(ev.Kv.Key), clientv3.WithPrefix())
+				if err != nil {
+					log.Println(err)
+					continue
+				}
+				// Try to decode the message
+				log.Println(string(ev.Kv.Value))
+				sexp, err := zygo.JsonToSexp(ev.Kv.Value, env)
+				_, err = lda.Eval(env, "eval", sexp)
+				if err != nil {
+					log.Println(err)
+				}
 			}
-			// Try to decode the message
-			buf := bytes.NewBuffer(ev.Kv.Value)
-			dec := gob.NewDecoder(buf)
-			err = dec.Decode(&msg)
-			if err != nil {
-				log.Fatal("decode error 1:", err)
-			}
-			break Loop
 		}
 	}
-	env := msg.Env
-	env.Run()
-	/*
-		fn := msg.Args[0].(*zygo.SexpFunction)
-		expr, err := env.Apply(fn, msg.Args[1:]) // Put the result in the tuplespace
-		if err != nil {
-			// TODO
-			log.Println(err)
-		}
-		if expr != zygo.SexpNull {
-			_, err := cli.Put(context.TODO(), prefix+"-"+uuid.New().String(), expr.SexpString(&zygo.PrintState{}))
-			log.Println(err)
-			return
-		}
-	*/
 	done := make(chan bool)
 	<-done
 	//zygo.Repl(env, cfg)
